Shims layer for HBase (support both hbase 0.98 and hbase 1.1.1)
