+++ hbase-client/pom.xml
@@ -141,7 +141,6 @@
         <dependency>
             <groupId>org.apache.hbase</groupId>
             <artifactId>hbase-client</artifactId>
-            <version>${hbase.version}</version>
             <exclusions>
               <exclusion>
                 <groupId>org.slf4j</groupId>
@@ -152,7 +151,6 @@
         <dependency>
             <groupId>org.apache.hbase</groupId>
             <artifactId>hbase-testing-util</artifactId>
-            <version>${hbase.version}</version>
             <scope>test</scope>
         </dependency>
         <dependency>
@@ -162,5 +160,33 @@
             <scope>test</scope>
         </dependency>
     </dependencies>
+
+  <profiles>
+
+    <profile>
+      <id>hbase-0</id>
+      <activation>
+        <activeByDefault>true</activeByDefault>
+      </activation>
+      <dependencies>
+        <dependency>
+        <groupId>com.yahoo.omid</groupId>
+          <artifactId>hbase0-shims</artifactId>
+          <version>${project.version}</version>
+        </dependency>
+      </dependencies>
+    </profile>
+
+    <profile>
+      <id>hbase-1</id>
+      <dependencies>
+        <dependency>
+        <groupId>com.yahoo.omid</groupId>
+          <artifactId>hbase1-shims</artifactId>
+          <version>${project.version}</version>
+        </dependency>
+      </dependencies>
+    </profile>
+  </profiles>
 
 </project>
