+++ hbase-shims/hbase-0/pom.xml
@@ -1,0 +1,21 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+
+  <modelVersion>4.0.0</modelVersion>
+
+  <parent>
+    <groupId>com.yahoo.omid</groupId>
+    <artifactId>omid-shims-aggregator</artifactId>
+    <version>0.8.0</version>
+  </parent>
+
+  <artifactId>hbase0-shims</artifactId>
+  <name>Shims layer for Hbase 0.x</name>
+  <packaging>jar</packaging>
+
+  <properties>
+    <hbase.version>${hbase0.version}</hbase.version>
+  </properties>
+
+</project>
