+++ hbase-shims/hbase-0/src/main/java/org/apache/hadoop/hbase/regionserver/ScannerContext.java
@@ -1,0 +1,7 @@
+package org.apache.hadoop.hbase.regionserver;
+
+public class ScannerContext {
+  int getBatchLimit() {
+    return -1;
+  }
+}
